image_classification/tensorflow2/requirements.txt
PyYAML==5.4.1
typer==0.3.2